refactor: enhance transaction handling with utility methods
- Added trait extension to make it more organized.
- Introduced `TransactionUtil` for transaction encoding/decoding and signing operations.
- Updated transaction methods to utilize the new utility functions for better organization and readability.
- Refactored transaction fee estimation and signing processes to leverage the new utility methods.
- Improved modularity by separating versioned message and transaction functionalities into dedicated files.
- Enhanced tests to ensure correctness of new utility methods and transaction handling.
