+++ crates/lib/src/transaction/mod.rs
@@ -1,3 +1,7 @@
 mod transaction;
+mod versioned_message;
+mod versioned_transaction;
 
 pub use transaction::*;
+pub use versioned_message::*;
+pub use versioned_transaction::*;
