+++ crates/lib/src/transaction/transaction.rs
@@ -1,8 +1,4 @@
-use std::ops::Deref;
-
-use solana_client::nonblocking::rpc_client::RpcClient;
-use solana_commitment_config::CommitmentConfig;
-use solana_message::{v0::MessageAddressTableLookup, VersionedMessage};
+use solana_message::VersionedMessage;
 use solana_sdk::{
     instruction::{AccountMeta, CompiledInstruction, Instruction},
     pubkey::Pubkey,
@@ -10,313 +6,64 @@ use solana_sdk::{
     transaction::{Transaction, VersionedTransaction},
 };
 
-use crate::{
-    config::ValidationConfig,
-    error::KoraError,
-    fee::fee::FeeConfigUtil,
-    get_signer,
-    validator::transaction_validator::{validate_token_payment, TransactionValidator},
-    Signer as _,
-};
+use crate::error::KoraError;
 use base64::{engine::general_purpose::STANDARD, Engine as _};
-use solana_address_lookup_table_interface::state::AddressLookupTable;
 
-pub trait VersionedTransactionExt {
-    fn get_all_account_keys(&self) -> Vec<Pubkey>;
-    fn get_transaction(&self) -> &VersionedTransaction;
-}
+pub struct TransactionUtil {}
 
-pub struct VersionedTransactionResolved<'a> {
-    pub transaction: &'a VersionedTransaction,
+impl TransactionUtil {
+    pub fn decode_b64_transaction(encoded: &str) -> Result<VersionedTransaction, KoraError> {
+        let decoded = STANDARD.decode(encoded).map_err(|e| {
+            KoraError::InvalidTransaction(format!("Failed to decode base64 transaction: {e}"))
+        })?;
 
-    pub is_resolved: bool,
-
-    pub resolved_addresses: Vec<Pubkey>,
-}
-
-impl Deref for VersionedTransactionResolved<'_> {
-    type Target = VersionedTransaction;
-
-    fn deref(&self) -> &Self::Target {
-        self.transaction
-    }
-}
-
-impl<'a> VersionedTransactionExt for VersionedTransactionResolved<'a> {
-    fn get_all_account_keys(&self) -> Vec<Pubkey> {
-        let mut account_keys = self.transaction.message.static_account_keys().to_vec();
-        account_keys.append(&mut self.resolved_addresses.clone());
-        account_keys
-    }
-
-    fn get_transaction(&self) -> &VersionedTransaction {
-        self.transaction
-    }
-}
-
-impl<'a> VersionedTransactionResolved<'a> {
-    pub fn new(transaction: &'a VersionedTransaction) -> Self {
-        Self { transaction, is_resolved: false, resolved_addresses: vec![] }
-    }
-
-    pub async fn resolve_addresses(&mut self, rpc_client: &RpcClient) -> Result<(), KoraError> {
-        if self.is_resolved {
-            return Ok(());
+        // First try to deserialize as VersionedTransaction
+        if let Ok(versioned_tx) = bincode::deserialize::<VersionedTransaction>(&decoded) {
+            return Ok(versioned_tx);
         }
 
-        self.resolved_addresses = match &self.transaction.message {
-            VersionedMessage::Legacy(_) => vec![],
-            VersionedMessage::V0(v0_message) => {
-                resolve_lookup_table_addresses(rpc_client, &v0_message.address_table_lookups)
-                    .await?
-            }
-        };
+        // Fall back to legacy Transaction and convert to VersionedTransaction
+        let legacy_tx: Transaction = bincode::deserialize(&decoded).map_err(|e| {
+            KoraError::InvalidTransaction(format!("Failed to deserialize transaction: {e}"))
+        })?;
 
-        Ok(())
-    }
-}
-
-impl VersionedTransactionExt for VersionedTransaction {
-    fn get_all_account_keys(&self) -> Vec<Pubkey> {
-        self.message.static_account_keys().to_vec()
-    }
-
-    fn get_transaction(&self) -> &VersionedTransaction {
-        self
-    }
-}
-
-pub fn new_unsigned_versioned_transaction(message: VersionedMessage) -> VersionedTransaction {
-    let num_required_signatures = message.header().num_required_signatures as usize;
-    VersionedTransaction {
-        signatures: vec![Signature::default(); num_required_signatures],
-        message,
-    }
-}
-
-pub async fn get_estimate_fee(
-    rpc_client: &RpcClient,
-    message: &VersionedMessage,
-) -> Result<u64, KoraError> {
-    match message {
-        VersionedMessage::Legacy(message) => rpc_client.get_fee_for_message(message).await,
-        VersionedMessage::V0(message) => rpc_client.get_fee_for_message(message).await,
-    }
-    .map_err(|e| KoraError::RpcError(e.to_string()))
-}
-
-pub fn uncompile_instructions(
-    instructions: &[CompiledInstruction],
-    account_keys: &[Pubkey],
-) -> Vec<Instruction> {
-    instructions
-        .iter()
-        .map(|ix| {
-            let program_id = account_keys[ix.program_id_index as usize];
-            let accounts = ix
-                .accounts
-                .iter()
-                .map(|idx| AccountMeta {
-                    pubkey: account_keys[*idx as usize],
-                    is_signer: false,
-                    is_writable: true,
-                })
-                .collect();
-
-            Instruction { program_id, accounts, data: ix.data.clone() }
+        // Convert legacy Transaction to VersionedTransaction
+        Ok(VersionedTransaction {
+            signatures: legacy_tx.signatures,
+            message: VersionedMessage::Legacy(legacy_tx.message),
         })
-        .collect()
-}
-
-pub async fn sign_transaction(
-    rpc_client: &RpcClient,
-    validation: &ValidationConfig,
-    transaction: VersionedTransaction,
-) -> Result<(VersionedTransaction, String), KoraError> {
-    let signer = get_signer()?;
-    let validator = TransactionValidator::new(signer.solana_pubkey(), validation)?;
-
-    let mut resolved_transaction = VersionedTransactionResolved::new(&transaction);
-    resolved_transaction.resolve_addresses(rpc_client).await?;
-
-    // Validate transaction and accounts with lookup table resolution for V0 transactions
-    validator.validate_transaction(rpc_client, &resolved_transaction).await?;
-
-    // Get latest blockhash and update transaction
-    let mut transaction = resolved_transaction.get_transaction().clone();
-
-    if transaction.signatures.is_empty() {
-        let blockhash =
-            rpc_client.get_latest_blockhash_with_commitment(CommitmentConfig::finalized()).await?;
-        transaction.message.set_recent_blockhash(blockhash.0);
     }
 
-    // Validate transaction fee
-    let estimated_fee = get_estimate_fee(rpc_client, &transaction.message).await?;
-    validator.validate_lamport_fee(estimated_fee)?;
-
-    // Sign transaction
-    let signature = signer.sign_solana(&transaction).await?;
-    transaction.signatures[0] = signature;
-
-    // Serialize signed transaction
-    let serialized = bincode::serialize(&transaction)?;
-    let encoded = STANDARD.encode(serialized);
-
-    Ok((transaction, encoded))
-}
-
-pub async fn sign_transaction_if_paid(
-    rpc_client: &RpcClient,
-    validation: &ValidationConfig,
-    // Should have resolved addresses for lookup tables
-    resolved_transaction: &impl VersionedTransactionExt,
-) -> Result<(VersionedTransaction, String), KoraError> {
-    let signer = get_signer()?;
-    let fee_payer = signer.solana_pubkey();
-
-    // Get the simulation result for fee calculation
-    let min_transaction_fee =
-        FeeConfigUtil::estimate_transaction_fee(rpc_client, resolved_transaction, Some(&fee_payer))
-            .await?;
-
-    let required_lamports = validation
-        .price
-        .get_required_lamports(
-            Some(rpc_client),
-            Some(validation.price_source.clone()),
-            min_transaction_fee,
-        )
-        .await?;
-
-    // Only validate payment if not free
-    if required_lamports > 0 {
-        // Validate token payment
-        validate_token_payment(
-            resolved_transaction,
-            required_lamports,
-            validation,
-            rpc_client,
-            signer.solana_pubkey(),
-        )
-        .await?;
-    }
-
-    let transaction = resolved_transaction.get_transaction().clone();
-
-    // Sign the transaction
-    sign_transaction(rpc_client, validation, transaction).await
-}
-
-pub async fn sign_and_send_transaction(
-    rpc_client: &RpcClient,
-    validation: &ValidationConfig,
-    transaction: VersionedTransaction,
-) -> Result<(String, String), KoraError> {
-    let (transaction, encoded) = sign_transaction(rpc_client, validation, transaction).await?;
-
-    // Send and confirm transaction
-    let signature = rpc_client
-        .send_and_confirm_transaction(&transaction)
-        .await
-        .map_err(|e| KoraError::RpcError(e.to_string()))?;
-
-    Ok((signature.to_string(), encoded))
-}
-
-pub fn encode_b64_transaction(transaction: &VersionedTransaction) -> Result<String, KoraError> {
-    let serialized = bincode::serialize(transaction)
-        .map_err(|e| KoraError::SerializationError(format!("Base64 serialization failed: {e}")))?;
-    Ok(STANDARD.encode(serialized))
-}
-
-pub fn encode_b64_message(message: &VersionedMessage) -> Result<String, KoraError> {
-    let serialized = message.serialize();
-    Ok(STANDARD.encode(serialized))
-}
-
-pub fn decode_b64_transaction(encoded: &str) -> Result<VersionedTransaction, KoraError> {
-    let decoded = STANDARD.decode(encoded).map_err(|e| {
-        KoraError::InvalidTransaction(format!("Failed to decode base64 transaction: {e}"))
-    })?;
-
-    // First try to deserialize as VersionedTransaction
-    if let Ok(versioned_tx) = bincode::deserialize::<VersionedTransaction>(&decoded) {
-        return Ok(versioned_tx);
-    }
-
-    // Fall back to legacy Transaction and convert to VersionedTransaction
-    let legacy_tx: Transaction = bincode::deserialize(&decoded).map_err(|e| {
-        KoraError::InvalidTransaction(format!("Failed to deserialize transaction: {e}"))
-    })?;
-
-    // Convert legacy Transaction to VersionedTransaction
-    Ok(VersionedTransaction {
-        signatures: legacy_tx.signatures,
-        message: VersionedMessage::Legacy(legacy_tx.message),
-    })
-}
-
-/// Resolves addresses from lookup tables for V0 transactions
-pub async fn resolve_lookup_table_addresses(
-    rpc_client: &RpcClient,
-    lookup_table_lookups: &[MessageAddressTableLookup],
-) -> Result<Vec<Pubkey>, KoraError> {
-    let mut resolved_addresses = Vec::new();
-
-    // Maybe we can use caching here, there's a chance the lookup tables get updated though, so tbd
-    for lookup in lookup_table_lookups {
-        let lookup_table_account = rpc_client
-            .get_account(&lookup.account_key)
-            .await
-            .map_err(|e| KoraError::RpcError(format!("Failed to fetch lookup table: {e}")))?;
-
-        // Parse the lookup table account data to get the actual addresses
-        let address_lookup_table = AddressLookupTable::deserialize(&lookup_table_account.data)
-            .map_err(|e| {
-                KoraError::InvalidTransaction(format!("Failed to deserialize lookup table: {e}"))
-            })?;
-
-        // Resolve writable addresses
-        for &index in &lookup.writable_indexes {
-            if let Some(address) = address_lookup_table.addresses.get(index as usize) {
-                resolved_addresses.push(*address);
-            } else {
-                return Err(KoraError::InvalidTransaction(format!(
-                    "Lookup table index {index} out of bounds for writable addresses"
-                )));
-            }
-        }
-
-        // Resolve readonly addresses
-        for &index in &lookup.readonly_indexes {
-            if let Some(address) = address_lookup_table.addresses.get(index as usize) {
-                resolved_addresses.push(*address);
-            } else {
-                return Err(KoraError::InvalidTransaction(format!(
-                    "Lookup table index {index} out of bounds for readonly addresses"
-                )));
-            }
+    pub fn new_unsigned_versioned_transaction(message: VersionedMessage) -> VersionedTransaction {
+        let num_required_signatures = message.header().num_required_signatures as usize;
+        VersionedTransaction {
+            signatures: vec![Signature::default(); num_required_signatures],
+            message,
         }
     }
 
-    Ok(resolved_addresses)
-}
+    pub fn uncompile_instructions(
+        instructions: &[CompiledInstruction],
+        account_keys: &[Pubkey],
+    ) -> Vec<Instruction> {
+        instructions
+            .iter()
+            .map(|ix| {
+                let program_id = account_keys[ix.program_id_index as usize];
+                let accounts = ix
+                    .accounts
+                    .iter()
+                    .map(|idx| AccountMeta {
+                        pubkey: account_keys[*idx as usize],
+                        is_signer: false,
+                        is_writable: true,
+                    })
+                    .collect();
 
-pub fn find_signer_position(
-    transaction: &VersionedTransaction,
-    signer_pubkey: &Pubkey,
-) -> Result<usize, KoraError> {
-    transaction
-        .message
-        .static_account_keys()
-        .iter()
-        .position(|key| key == signer_pubkey)
-        .ok_or_else(|| {
-            KoraError::InvalidTransaction(format!(
-                "Signer {signer_pubkey} not found in transaction account keys"
-            ))
-        })
+                Instruction { program_id, accounts, data: ix.data.clone() }
+            })
+            .collect()
+    }
 }
 
 #[cfg(test)]
@@ -326,13 +73,9 @@ mod tests {
     use super::*;
     use crate::error::KoraError;
     use serde_json::json;
-    use solana_address_lookup_table_interface::state::LookupTableMeta;
-    use solana_client::rpc_request::RpcRequest;
+    use solana_client::{nonblocking::rpc_client::RpcClient, rpc_request::RpcRequest};
     use solana_message::{v0, Message};
-    use solana_sdk::{
-        account::Account, hash::Hash, signature::Keypair, signer::Signer as _,
-        transaction::VersionedTransaction,
-    };
+    use solana_sdk::{account::Account, hash::Hash, signature::Keypair, signer::Signer as _};
 
     fn get_mock_rpc_client(account: &Account) -> Arc<RpcClient> {
         let mut mocks = HashMap::new();
@@ -356,49 +99,12 @@ mod tests {
     }
 
     #[test]
-    fn test_encode_decode_b64_transaction() {
-        let keypair = Keypair::new();
-        let instruction = Instruction::new_with_bytes(
-            Pubkey::new_unique(),
-            &[1, 2, 3],
-            vec![AccountMeta::new(keypair.pubkey(), true)],
-        );
-        let message =
-            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
-        let tx = VersionedTransaction::try_new(message, &[&keypair]).unwrap();
-
-        let encoded = encode_b64_transaction(&tx).unwrap();
-        let decoded = decode_b64_transaction(&encoded).unwrap();
-
-        assert_eq!(tx, decoded);
-    }
-
-    #[test]
     fn test_decode_b64_transaction_invalid_input() {
-        let result = decode_b64_transaction("not-base64!");
+        let result = TransactionUtil::decode_b64_transaction("not-base64!");
         assert!(matches!(result, Err(KoraError::InvalidTransaction(_))));
 
-        let result = decode_b64_transaction("AQID"); // base64 of [1,2,3]
+        let result = TransactionUtil::decode_b64_transaction("AQID"); // base64 of [1,2,3]
         assert!(matches!(result, Err(KoraError::InvalidTransaction(_))));
-    }
-
-    #[test]
-    fn test_encode_transaction_b64() {
-        let keypair = Keypair::new();
-        let instruction = Instruction::new_with_bytes(
-            Pubkey::new_unique(),
-            &[1, 2, 3],
-            vec![AccountMeta::new(keypair.pubkey(), true)],
-        );
-        let message =
-            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
-        let tx = VersionedTransaction::try_new(message, &[&keypair]).unwrap();
-
-        let encoded = encode_b64_transaction(&tx).unwrap();
-        assert!(!encoded.is_empty());
-        assert!(encoded
-            .chars()
-            .all(|c| c.is_ascii_alphanumeric() || c == '+' || c == '/' || c == '='));
     }
 
     #[test]
@@ -414,7 +120,7 @@ mod tests {
             data: vec![1, 2, 3],
         };
 
-        let instructions = uncompile_instructions(&[compiled_ix], &account_keys);
+        let instructions = TransactionUtil::uncompile_instructions(&[compiled_ix], &account_keys);
 
         assert_eq!(instructions.len(), 1);
         let uncompiled = &instructions[0];
@@ -436,7 +142,7 @@ mod tests {
         let message =
             VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
 
-        let transaction = new_unsigned_versioned_transaction(message.clone());
+        let transaction = TransactionUtil::new_unsigned_versioned_transaction(message.clone());
 
         // Should have correct number of signatures (all default/empty)
         assert_eq!(transaction.signatures.len(), message.header().num_required_signatures as usize);
@@ -474,7 +180,7 @@ mod tests {
         };
         let message = VersionedMessage::V0(v0_message);
 
-        let transaction = new_unsigned_versioned_transaction(message.clone());
+        let transaction = TransactionUtil::new_unsigned_versioned_transaction(message.clone());
 
         assert_eq!(transaction.signatures.len(), 1);
         assert_eq!(transaction.signatures[0], Signature::default());
@@ -497,7 +203,7 @@ mod tests {
         let serialized = bincode::serialize(&legacy_tx).unwrap();
         let encoded = base64::engine::general_purpose::STANDARD.encode(serialized);
 
-        let decoded = decode_b64_transaction(&encoded).unwrap();
+        let decoded = TransactionUtil::decode_b64_transaction(&encoded).unwrap();
 
         match decoded.message {
             VersionedMessage::Legacy(msg) => {
@@ -505,308 +211,6 @@ mod tests {
                 assert_eq!(msg.account_keys.len(), 2); // keypair + program_id
             }
             VersionedMessage::V0(_) => panic!("Expected legacy message after conversion"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_resolve_lookup_table_addresses() {
-        let lookup_account_key = Pubkey::new_unique();
-        let address1 = Pubkey::new_unique();
-        let address2 = Pubkey::new_unique();
-        let address3 = Pubkey::new_unique();
-
-        let lookup_table = AddressLookupTable {
-            meta: LookupTableMeta {
-                deactivation_slot: u64::MAX,
-                last_extended_slot: 0,
-                last_extended_slot_start_index: 0,
-                authority: Some(Pubkey::new_unique()),
-                _padding: 0,
-            },
-            addresses: vec![address1, address2, address3].into(),
-        };
-
-        let serialized_data = lookup_table.serialize_for_tests().unwrap();
-
-        let rpc_client = get_mock_rpc_client(&Account {
-            data: serialized_data,
-            executable: false,
-            lamports: 0,
-            owner: Pubkey::new_unique(),
-            rent_epoch: 0,
-        });
-
-        let lookups = vec![solana_message::v0::MessageAddressTableLookup {
-            account_key: lookup_account_key,
-            writable_indexes: vec![0, 2], // address1, address3
-            readonly_indexes: vec![1],    // address2
-        }];
-
-        let resolved_addresses =
-            resolve_lookup_table_addresses(&rpc_client, &lookups).await.unwrap();
-
-        assert_eq!(resolved_addresses.len(), 3);
-        assert_eq!(resolved_addresses[0], address1);
-        assert_eq!(resolved_addresses[1], address3);
-        assert_eq!(resolved_addresses[2], address2);
-    }
-
-    #[test]
-    fn test_find_signer_position_success() {
-        let keypair = Keypair::new();
-        let program_id = Pubkey::new_unique();
-        let instruction = Instruction::new_with_bytes(
-            program_id,
-            &[1, 2, 3],
-            vec![AccountMeta::new(keypair.pubkey(), true)],
-        );
-        let message =
-            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
-        let transaction = new_unsigned_versioned_transaction(message);
-
-        let position = find_signer_position(&transaction, &keypair.pubkey()).unwrap();
-        assert_eq!(position, 0); // Fee payer is typically at position 0
-    }
-
-    #[test]
-    fn test_find_signer_position_success_v0() {
-        let keypair = Keypair::new();
-        let program_id = Pubkey::new_unique();
-        let other_account = Pubkey::new_unique();
-
-        let v0_message = v0::Message {
-            header: solana_message::MessageHeader {
-                num_required_signatures: 1,
-                num_readonly_signed_accounts: 0,
-                num_readonly_unsigned_accounts: 2,
-            },
-            account_keys: vec![keypair.pubkey(), other_account, program_id],
-            recent_blockhash: Hash::default(),
-            instructions: vec![CompiledInstruction {
-                program_id_index: 2,
-                accounts: vec![0, 1],
-                data: vec![1, 2, 3],
-            }],
-            address_table_lookups: vec![],
-        };
-        let message = VersionedMessage::V0(v0_message);
-        let transaction = new_unsigned_versioned_transaction(message);
-
-        let position = find_signer_position(&transaction, &keypair.pubkey()).unwrap();
-        assert_eq!(position, 0);
-
-        let other_position = find_signer_position(&transaction, &other_account).unwrap();
-        assert_eq!(other_position, 1);
-    }
-
-    #[test]
-    fn test_find_signer_position_not_found() {
-        let keypair = Keypair::new();
-        let missing_keypair = Keypair::new();
-        let instruction = Instruction::new_with_bytes(
-            Pubkey::new_unique(),
-            &[1, 2, 3],
-            vec![AccountMeta::new(keypair.pubkey(), true)],
-        );
-        let message =
-            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
-        let transaction = new_unsigned_versioned_transaction(message);
-
-        let result = find_signer_position(&transaction, &missing_keypair.pubkey());
-        assert!(matches!(result, Err(KoraError::InvalidTransaction(_))));
-
-        if let Err(KoraError::InvalidTransaction(msg)) = result {
-            assert!(msg.contains(&missing_keypair.pubkey().to_string()));
-            assert!(msg.contains("not found in transaction account keys"));
-        }
-    }
-
-    #[test]
-    fn test_find_signer_position_empty_account_keys() {
-        // Create a transaction with minimal account keys
-        let v0_message = v0::Message {
-            header: solana_message::MessageHeader {
-                num_required_signatures: 0,
-                num_readonly_signed_accounts: 0,
-                num_readonly_unsigned_accounts: 0,
-            },
-            account_keys: vec![], // Empty account keys
-            recent_blockhash: Hash::default(),
-            instructions: vec![],
-            address_table_lookups: vec![],
-        };
-        let message = VersionedMessage::V0(v0_message);
-        let transaction = new_unsigned_versioned_transaction(message);
-        let search_key = Pubkey::new_unique();
-
-        let result = find_signer_position(&transaction, &search_key);
-        assert!(matches!(result, Err(KoraError::InvalidTransaction(_))));
-    }
-
-    #[cfg(test)]
-    mod tests {
-
-        use std::{collections::HashMap, sync::Arc};
-
-        use base64::{engine::general_purpose::STANDARD, Engine as _};
-        use serde_json::json;
-        use solana_client::{nonblocking::rpc_client::RpcClient, rpc_request::RpcRequest};
-        use solana_message::{Message, VersionedMessage};
-        use solana_sdk::{
-            hash::Hash,
-            instruction::{AccountMeta, Instruction},
-            pubkey::Pubkey,
-            signature::Keypair,
-            signer::Signer as _,
-            transaction::{Transaction, VersionedTransaction},
-        };
-        use solana_system_interface::instruction::transfer;
-
-        use crate::{
-            fee::fee::FeeConfigUtil,
-            get_signer, init_signer,
-            signer::{KoraSigner, SolanaMemorySigner},
-            token::{token::TokenType, TokenInterface, TokenProgram},
-            transaction::decode_b64_transaction,
-        };
-
-        fn setup_test_rpc_client() -> Arc<RpcClient> {
-            let rpc_url = "http://localhost:8899".to_string();
-            let mut mocks = HashMap::new();
-            mocks.insert(RpcRequest::GetMinimumBalanceForRentExemption, json!(2_039_280));
-            mocks.insert(
-                RpcRequest::GetFeeForMessage,
-                json!({
-                    "context": { "slot": 5068 },
-                    "value": 5000
-                }),
-            ); // Mock fee of 5000 lamports
-            Arc::new(RpcClient::new_mock_with_mocks(rpc_url, mocks))
-        }
-
-        fn ensure_test_signer_initialized() {
-            if get_signer().is_err() {
-                let test_keypair = Keypair::new();
-                let signer = SolanaMemorySigner::new(test_keypair);
-                let _ = init_signer(KoraSigner::Memory(signer));
-            }
-        }
-
-        #[test]
-        fn test_decode_b64_transaction() {
-            let keypair = Keypair::new();
-            let instruction = Instruction::new_with_bytes(
-                Pubkey::new_unique(),
-                &[1, 2, 3],
-                vec![AccountMeta::new(keypair.pubkey(), true)],
-            );
-            let message = Message::new(&[instruction], Some(&keypair.pubkey()));
-            let legacy_tx = Transaction::new(&[&keypair], message, Hash::default());
-
-            // Convert legacy transaction to versioned transaction for encoding
-            let tx = VersionedTransaction {
-                signatures: legacy_tx.signatures,
-                message: VersionedMessage::Legacy(legacy_tx.message),
-            };
-
-            let encoded = STANDARD.encode(bincode::serialize(&tx).unwrap());
-            let decoded = decode_b64_transaction(&encoded).unwrap();
-
-            assert_eq!(tx, decoded);
-        }
-
-        #[test]
-        fn test_decode_b64_transaction_invalid_input() {
-            let result = decode_b64_transaction("not-base64!");
-            assert!(matches!(result, Err(crate::error::KoraError::InvalidTransaction(_))));
-
-            let result = decode_b64_transaction("AQID"); // base64 of [1,2,3]
-            assert!(matches!(result, Err(crate::error::KoraError::InvalidTransaction(_))));
-        }
-
-        #[tokio::test]
-        async fn test_estimate_transaction_fee_basic() {
-            ensure_test_signer_initialized();
-            let rpc_client = setup_test_rpc_client();
-
-            // Create a simple transfer transaction
-            let from = Pubkey::new_unique();
-            let to = Pubkey::new_unique();
-            let instruction = transfer(&from, &to, 1000);
-            let message = Message::new(&[instruction], Some(&from));
-            let transaction = VersionedTransaction {
-                signatures: vec![Default::default()],
-                message: VersionedMessage::Legacy(message),
-            };
-
-            let fee = FeeConfigUtil::estimate_transaction_fee(&rpc_client, &transaction, None)
-                .await
-                .unwrap();
-
-            // Base fee + priority fee
-            assert!(fee > 0);
-            // Fee should be less than the minimum rent-exempt amount for a token account (~0.00204 SOL)
-            assert!(fee < 2_039_280);
-        }
-
-        #[tokio::test]
-        async fn test_estimate_transaction_fee_invalid_transaction() {
-            ensure_test_signer_initialized();
-            let rpc_client = setup_test_rpc_client();
-
-            // Create an invalid transaction (empty message)
-            let transaction = VersionedTransaction {
-                signatures: vec![],
-                message: VersionedMessage::Legacy(Message::default()),
-            };
-
-            let result =
-                FeeConfigUtil::estimate_transaction_fee(&rpc_client, &transaction, None).await;
-            assert!(result.is_ok()); // Fee estimation should still work for invalid transactions
-        }
-
-        #[tokio::test]
-        async fn test_estimate_transaction_fee_with_token_creation() {
-            ensure_test_signer_initialized();
-            let rpc_client = setup_test_rpc_client();
-
-            // Create a transaction that includes token account creation
-            let payer = Pubkey::new_unique();
-            let mint = Pubkey::new_unique();
-            let owner = Pubkey::new_unique();
-
-            let token_program = TokenProgram::new(TokenType::Spl);
-            let _ata = token_program.get_associated_token_address(&owner, &mint);
-            let create_ata_ix =
-                token_program.create_associated_token_account_instruction(&payer, &owner, &mint);
-
-            let message = Message::new(&[create_ata_ix], Some(&payer));
-            let transaction = VersionedTransaction {
-                signatures: vec![Default::default()],
-                message: VersionedMessage::Legacy(message),
-            };
-
-            let fee = FeeConfigUtil::estimate_transaction_fee(&rpc_client, &transaction, None)
-                .await
-                .unwrap();
-
-            // Fee should include base fee + priority fee + rent for token account
-            let min_expected_lamports = 2_039_280;
-            assert!(
-                fee >= min_expected_lamports,
-                "Fee {fee} lamports is less than minimum expected {min_expected_lamports} lamports"
-            );
-        }
-
-        #[test]
-        fn test_token_functionality() {
-            let token_program = TokenProgram::new(TokenType::Spl);
-            let mint = Pubkey::new_unique();
-            let owner = Pubkey::new_unique();
-
-            let ata = token_program.get_associated_token_address(&owner, &mint);
-            assert_ne!(ata, owner);
-            assert_ne!(ata, mint);
         }
     }
 }
