+++ crates/lib/src/transaction/versioned_transaction.rs
@@ -1,0 +1,488 @@
+use async_trait::async_trait;
+use base64::{engine::general_purpose::STANDARD, Engine as _};
+use solana_client::nonblocking::rpc_client::RpcClient;
+use solana_commitment_config::CommitmentConfig;
+use solana_message::{v0::MessageAddressTableLookup, VersionedMessage};
+use solana_sdk::{pubkey::Pubkey, transaction::VersionedTransaction};
+use std::ops::Deref;
+
+use crate::{
+    config::ValidationConfig,
+    error::KoraError,
+    fee::fee::{FeeConfigUtil, TransactionFeeUtil},
+    get_signer,
+    validator::transaction_validator::{validate_token_payment, TransactionValidator},
+    Signer,
+};
+use solana_address_lookup_table_interface::state::AddressLookupTable;
+
+#[async_trait]
+pub trait VersionedTransactionExt: Sized {
+    fn get_all_account_keys(&self) -> Vec<Pubkey>;
+    fn get_transaction(&self) -> &VersionedTransaction;
+
+    async fn sign_transaction_if_paid(
+        // Should have resolved addresses for lookup tables
+        &self,
+        rpc_client: &RpcClient,
+        validation: &ValidationConfig,
+    ) -> Result<(VersionedTransaction, String), KoraError> {
+        let signer = get_signer()?;
+        let fee_payer = signer.solana_pubkey();
+
+        // Get the simulation result for fee calculation
+        let min_transaction_fee =
+            FeeConfigUtil::estimate_transaction_fee(rpc_client, self, Some(&fee_payer)).await?;
+
+        let required_lamports = validation
+            .price
+            .get_required_lamports(
+                Some(rpc_client),
+                Some(validation.price_source.clone()),
+                min_transaction_fee,
+            )
+            .await?;
+
+        // Only validate payment if not free
+        if required_lamports > 0 {
+            // Validate token payment
+            validate_token_payment(
+                self,
+                required_lamports,
+                validation,
+                rpc_client,
+                signer.solana_pubkey(),
+            )
+            .await?;
+        }
+
+        let transaction = self.get_transaction().clone();
+
+        // Sign the transaction
+        transaction.sign_transaction(rpc_client, validation).await
+    }
+}
+
+pub struct VersionedTransactionResolved<'a> {
+    pub transaction: &'a VersionedTransaction,
+
+    pub is_resolved: bool,
+
+    pub resolved_addresses: Vec<Pubkey>,
+}
+
+impl Deref for VersionedTransactionResolved<'_> {
+    type Target = VersionedTransaction;
+
+    fn deref(&self) -> &Self::Target {
+        self.transaction
+    }
+}
+
+impl<'a> VersionedTransactionExt for VersionedTransactionResolved<'a> {
+    fn get_all_account_keys(&self) -> Vec<Pubkey> {
+        let mut account_keys = self.transaction.message.static_account_keys().to_vec();
+        account_keys.append(&mut self.resolved_addresses.clone());
+        account_keys
+    }
+
+    fn get_transaction(&self) -> &VersionedTransaction {
+        self.transaction
+    }
+}
+
+impl<'a> VersionedTransactionResolved<'a> {
+    pub fn new(transaction: &'a VersionedTransaction) -> Self {
+        Self { transaction, is_resolved: false, resolved_addresses: vec![] }
+    }
+
+    pub async fn resolve_addresses(&mut self, rpc_client: &RpcClient) -> Result<(), KoraError> {
+        if self.is_resolved {
+            return Ok(());
+        }
+
+        self.resolved_addresses = match &self.transaction.message {
+            VersionedMessage::Legacy(_) => vec![],
+            VersionedMessage::V0(v0_message) => {
+                LookupTableUtil::resolve_lookup_table_addresses(
+                    rpc_client,
+                    &v0_message.address_table_lookups,
+                )
+                .await?
+            }
+        };
+
+        Ok(())
+    }
+}
+
+impl VersionedTransactionExt for VersionedTransaction {
+    fn get_all_account_keys(&self) -> Vec<Pubkey> {
+        self.message.static_account_keys().to_vec()
+    }
+
+    fn get_transaction(&self) -> &VersionedTransaction {
+        self
+    }
+}
+
+#[async_trait]
+pub trait VersionedTransactionUtilExt {
+    fn encode_b64_transaction(&self) -> Result<String, KoraError>;
+
+    async fn sign_transaction(
+        &self,
+        rpc_client: &RpcClient,
+        validation: &ValidationConfig,
+    ) -> Result<(VersionedTransaction, String), KoraError>;
+
+    async fn sign_and_send_transaction(
+        &self,
+        rpc_client: &RpcClient,
+        validation: &ValidationConfig,
+    ) -> Result<(String, String), KoraError>;
+
+    fn find_signer_position(&self, signer_pubkey: &Pubkey) -> Result<usize, KoraError>;
+}
+
+#[async_trait]
+impl VersionedTransactionUtilExt for VersionedTransaction {
+    fn encode_b64_transaction(&self) -> Result<String, KoraError> {
+        let serialized = bincode::serialize(self).map_err(|e| {
+            KoraError::SerializationError(format!("Base64 serialization failed: {e}"))
+        })?;
+        Ok(STANDARD.encode(serialized))
+    }
+
+    async fn sign_transaction(
+        &self,
+        rpc_client: &RpcClient,
+        validation: &ValidationConfig,
+    ) -> Result<(VersionedTransaction, String), KoraError> {
+        let signer = get_signer()?;
+        let validator = TransactionValidator::new(signer.solana_pubkey(), validation)?;
+
+        let mut resolved_transaction = VersionedTransactionResolved::new(self);
+        resolved_transaction.resolve_addresses(rpc_client).await?;
+
+        // Validate transaction and accounts with lookup table resolution for V0 transactions
+        validator.validate_transaction(rpc_client, &resolved_transaction).await?;
+
+        // Get latest blockhash and update transaction
+        let mut transaction = resolved_transaction.get_transaction().clone();
+
+        if transaction.signatures.is_empty() {
+            let blockhash = rpc_client
+                .get_latest_blockhash_with_commitment(CommitmentConfig::finalized())
+                .await?;
+            transaction.message.set_recent_blockhash(blockhash.0);
+        }
+
+        // Validate transaction fee
+        let estimated_fee =
+            TransactionFeeUtil::get_estimate_fee(rpc_client, &transaction.message).await?;
+        validator.validate_lamport_fee(estimated_fee)?;
+
+        // Sign transaction
+        let signature = signer.sign_solana(&transaction).await?;
+        transaction.signatures[0] = signature;
+
+        // Serialize signed transaction
+        let serialized = bincode::serialize(&transaction)?;
+        let encoded = STANDARD.encode(serialized);
+
+        Ok((transaction, encoded))
+    }
+
+    async fn sign_and_send_transaction(
+        &self,
+        rpc_client: &RpcClient,
+        validation: &ValidationConfig,
+    ) -> Result<(String, String), KoraError> {
+        let (transaction, encoded) = self.sign_transaction(rpc_client, validation).await?;
+
+        // Send and confirm transaction
+        let signature = rpc_client
+            .send_and_confirm_transaction(&transaction)
+            .await
+            .map_err(|e| KoraError::RpcError(e.to_string()))?;
+
+        Ok((signature.to_string(), encoded))
+    }
+
+    fn find_signer_position(&self, signer_pubkey: &Pubkey) -> Result<usize, KoraError> {
+        self.message.static_account_keys().iter().position(|key| key == signer_pubkey).ok_or_else(
+            || {
+                KoraError::InvalidTransaction(format!(
+                    "Signer {signer_pubkey} not found in transaction account keys"
+                ))
+            },
+        )
+    }
+}
+
+pub struct LookupTableUtil {}
+
+impl LookupTableUtil {
+    /// Resolves addresses from lookup tables for V0 transactions
+    pub async fn resolve_lookup_table_addresses(
+        rpc_client: &RpcClient,
+        lookup_table_lookups: &[MessageAddressTableLookup],
+    ) -> Result<Vec<Pubkey>, KoraError> {
+        let mut resolved_addresses = Vec::new();
+
+        // Maybe we can use caching here, there's a chance the lookup tables get updated though, so tbd
+        for lookup in lookup_table_lookups {
+            let lookup_table_account = rpc_client
+                .get_account(&lookup.account_key)
+                .await
+                .map_err(|e| KoraError::RpcError(format!("Failed to fetch lookup table: {e}")))?;
+
+            // Parse the lookup table account data to get the actual addresses
+            let address_lookup_table = AddressLookupTable::deserialize(&lookup_table_account.data)
+                .map_err(|e| {
+                    KoraError::InvalidTransaction(format!(
+                        "Failed to deserialize lookup table: {e}"
+                    ))
+                })?;
+
+            // Resolve writable addresses
+            for &index in &lookup.writable_indexes {
+                if let Some(address) = address_lookup_table.addresses.get(index as usize) {
+                    resolved_addresses.push(*address);
+                } else {
+                    return Err(KoraError::InvalidTransaction(format!(
+                        "Lookup table index {index} out of bounds for writable addresses"
+                    )));
+                }
+            }
+
+            // Resolve readonly addresses
+            for &index in &lookup.readonly_indexes {
+                if let Some(address) = address_lookup_table.addresses.get(index as usize) {
+                    resolved_addresses.push(*address);
+                } else {
+                    return Err(KoraError::InvalidTransaction(format!(
+                        "Lookup table index {index} out of bounds for readonly addresses"
+                    )));
+                }
+            }
+        }
+
+        Ok(resolved_addresses)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{collections::HashMap, sync::Arc};
+
+    use crate::transaction::TransactionUtil;
+
+    use super::*;
+    use base64::Engine;
+    use serde_json::json;
+    use solana_address_lookup_table_interface::state::LookupTableMeta;
+    use solana_client::rpc_request::RpcRequest;
+    use solana_message::{v0, Message};
+    use solana_sdk::{
+        account::Account,
+        hash::Hash,
+        instruction::{AccountMeta, CompiledInstruction, Instruction},
+        signature::Keypair,
+        signer::Signer,
+    };
+
+    fn get_mock_rpc_client(account: &Account) -> Arc<RpcClient> {
+        let mut mocks = HashMap::new();
+        let encoded_data = base64::engine::general_purpose::STANDARD.encode(&account.data);
+        mocks.insert(
+            RpcRequest::GetAccountInfo,
+            json!({
+                "context": {
+                    "slot": 1
+                },
+                "value": {
+                    "data": [encoded_data, "base64"],
+                    "executable": account.executable,
+                    "lamports": account.lamports,
+                    "owner": account.owner.to_string(),
+                    "rentEpoch": account.rent_epoch
+                }
+            }),
+        );
+        Arc::new(RpcClient::new_mock_with_mocks("http://localhost:8899".to_string(), mocks))
+    }
+
+    #[test]
+    fn test_encode_transaction_b64() {
+        let keypair = Keypair::new();
+        let instruction = Instruction::new_with_bytes(
+            Pubkey::new_unique(),
+            &[1, 2, 3],
+            vec![AccountMeta::new(keypair.pubkey(), true)],
+        );
+        let message =
+            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
+        let tx = VersionedTransaction::try_new(message, &[&keypair]).unwrap();
+
+        let encoded = tx.encode_b64_transaction().unwrap();
+        assert!(!encoded.is_empty());
+        assert!(encoded
+            .chars()
+            .all(|c| c.is_ascii_alphanumeric() || c == '+' || c == '/' || c == '='));
+    }
+
+    #[test]
+    fn test_encode_decode_b64_transaction() {
+        let keypair = Keypair::new();
+        let instruction = Instruction::new_with_bytes(
+            Pubkey::new_unique(),
+            &[1, 2, 3],
+            vec![AccountMeta::new(keypair.pubkey(), true)],
+        );
+        let message =
+            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
+        let tx = VersionedTransaction::try_new(message, &[&keypair]).unwrap();
+
+        let encoded = tx.encode_b64_transaction().unwrap();
+        let decoded = TransactionUtil::decode_b64_transaction(&encoded).unwrap();
+
+        assert_eq!(tx, decoded);
+    }
+
+    #[test]
+    fn test_find_signer_position_success() {
+        let keypair = Keypair::new();
+        let program_id = Pubkey::new_unique();
+        let instruction = Instruction::new_with_bytes(
+            program_id,
+            &[1, 2, 3],
+            vec![AccountMeta::new(keypair.pubkey(), true)],
+        );
+        let message =
+            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
+        let transaction = TransactionUtil::new_unsigned_versioned_transaction(message);
+
+        let position = transaction.find_signer_position(&keypair.pubkey()).unwrap();
+        assert_eq!(position, 0); // Fee payer is typically at position 0
+    }
+
+    #[test]
+    fn test_find_signer_position_success_v0() {
+        let keypair = Keypair::new();
+        let program_id = Pubkey::new_unique();
+        let other_account = Pubkey::new_unique();
+
+        let v0_message = v0::Message {
+            header: solana_message::MessageHeader {
+                num_required_signatures: 1,
+                num_readonly_signed_accounts: 0,
+                num_readonly_unsigned_accounts: 2,
+            },
+            account_keys: vec![keypair.pubkey(), other_account, program_id],
+            recent_blockhash: Hash::default(),
+            instructions: vec![CompiledInstruction {
+                program_id_index: 2,
+                accounts: vec![0, 1],
+                data: vec![1, 2, 3],
+            }],
+            address_table_lookups: vec![],
+        };
+        let message = VersionedMessage::V0(v0_message);
+        let transaction = TransactionUtil::new_unsigned_versioned_transaction(message);
+
+        let position = transaction.find_signer_position(&keypair.pubkey()).unwrap();
+        assert_eq!(position, 0);
+
+        let other_position = transaction.find_signer_position(&other_account).unwrap();
+        assert_eq!(other_position, 1);
+    }
+
+    #[test]
+    fn test_find_signer_position_not_found() {
+        let keypair = Keypair::new();
+        let missing_keypair = Keypair::new();
+        let instruction = Instruction::new_with_bytes(
+            Pubkey::new_unique(),
+            &[1, 2, 3],
+            vec![AccountMeta::new(keypair.pubkey(), true)],
+        );
+        let message =
+            VersionedMessage::Legacy(Message::new(&[instruction], Some(&keypair.pubkey())));
+        let transaction = TransactionUtil::new_unsigned_versioned_transaction(message);
+
+        let result = transaction.find_signer_position(&missing_keypair.pubkey());
+        assert!(matches!(result, Err(KoraError::InvalidTransaction(_))));
+
+        if let Err(KoraError::InvalidTransaction(msg)) = result {
+            assert!(msg.contains(&missing_keypair.pubkey().to_string()));
+            assert!(msg.contains("not found in transaction account keys"));
+        }
+    }
+
+    #[test]
+    fn test_find_signer_position_empty_account_keys() {
+        // Create a transaction with minimal account keys
+        let v0_message = v0::Message {
+            header: solana_message::MessageHeader {
+                num_required_signatures: 0,
+                num_readonly_signed_accounts: 0,
+                num_readonly_unsigned_accounts: 0,
+            },
+            account_keys: vec![], // Empty account keys
+            recent_blockhash: Hash::default(),
+            instructions: vec![],
+            address_table_lookups: vec![],
+        };
+        let message = VersionedMessage::V0(v0_message);
+        let transaction = TransactionUtil::new_unsigned_versioned_transaction(message);
+        let search_key = Pubkey::new_unique();
+
+        let result = transaction.find_signer_position(&search_key);
+        assert!(matches!(result, Err(KoraError::InvalidTransaction(_))));
+    }
+
+    #[tokio::test]
+    async fn test_resolve_lookup_table_addresses() {
+        let lookup_account_key = Pubkey::new_unique();
+        let address1 = Pubkey::new_unique();
+        let address2 = Pubkey::new_unique();
+        let address3 = Pubkey::new_unique();
+
+        let lookup_table = AddressLookupTable {
+            meta: LookupTableMeta {
+                deactivation_slot: u64::MAX,
+                last_extended_slot: 0,
+                last_extended_slot_start_index: 0,
+                authority: Some(Pubkey::new_unique()),
+                _padding: 0,
+            },
+            addresses: vec![address1, address2, address3].into(),
+        };
+
+        let serialized_data = lookup_table.serialize_for_tests().unwrap();
+
+        let rpc_client = get_mock_rpc_client(&Account {
+            data: serialized_data,
+            executable: false,
+            lamports: 0,
+            owner: Pubkey::new_unique(),
+            rent_epoch: 0,
+        });
+
+        let lookups = vec![solana_message::v0::MessageAddressTableLookup {
+            account_key: lookup_account_key,
+            writable_indexes: vec![0, 2], // address1, address3
+            readonly_indexes: vec![1],    // address2
+        }];
+
+        let resolved_addresses =
+            LookupTableUtil::resolve_lookup_table_addresses(&rpc_client, &lookups).await.unwrap();
+
+        assert_eq!(resolved_addresses.len(), 3);
+        assert_eq!(resolved_addresses[0], address1);
+        assert_eq!(resolved_addresses[1], address3);
+        assert_eq!(resolved_addresses[2], address2);
+    }
+}
